run threads test with and without `nthreads > 1`
